+++ ChangeLog.md
@@ -3,6 +3,7 @@ Unreleased
 
 * Windows builds now use `-D_SDL_main_h`. See https://github.com/haskell-game/sdl2/issues/139 for more discussion.
 * Support for event watching: `addEventWatch` and `delEventWatch`.
+* High-level structure for joystick device connection: `JoyDeviceConnection`.
 
 2.2.0
 =====